ITERATION_3_Solution/KisaanDostApp/requirements.txt
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
pymongo==4.2.0
Werkzeug==2.2.2
pyyaml
requests
SpeechRecognition
numpy
pandas
regex
rank_bm25
transformers
sentencepiece